+++ st2common/st2common/models/api/rule.py
@@ -17,6 +17,7 @@ import copy
 
 import six
 
+from st2common.constants.pack import DEFAULT_PACK_NAME
 from st2common.models.api.base import BaseAPI
 from st2common.models.api.trigger import TriggerAPI
 from st2common.models.api.tag import TagsHelper
@@ -162,7 +163,7 @@ class RuleAPI(BaseAPI):
 
         trigger = reference.get_str_resource_ref_from_model(trigger_db)
         criteria = dict(getattr(rule, 'criteria', {}))
-        pack = str(rule.pack)
+        pack = getattr(rule, 'pack', DEFAULT_PACK_NAME)
         ref = ResourceReference.to_string_reference(pack=pack, name=name)
 
         # Validate criteria