+++ st2common/st2common/persistence/execution.py
@@ -13,12 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from oslo_config import cfg
-
 from st2common import transport
 from st2common.models.db import MongoDBAccess
 from st2common.models.db.execution import ActionExecutionDB
 from st2common.persistence.base import Access
+from st2common.transport import utils as transport_utils
 
 
 class ActionExecution(Access):
@@ -33,5 +32,5 @@ class ActionExecution(Access):
     def _get_publisher(cls):
         if not cls.publisher:
             cls.publisher = transport.execution.ActionExecutionPublisher(
-                cfg.CONF.messaging.url)
+                urls=transport_utils.get_messaging_urls())
         return cls.publisher