+++ st2common/tests/unit/test_db_liveaction.py
@@ -77,7 +77,7 @@ class LiveActionModelTest(DbTestCase):
         # Assert notify settings saved are right.
         self.assertEqual(notify_sub_schema.message, retrieved.notify.on_complete.message)
         self.assertDictEqual(notify_sub_schema.data, retrieved.notify.on_complete.data)
-        self.assertListEqual(notify_sub_schema.channels, retrieved.notify.on_complete.channels)
+        self.assertListEqual(notify_sub_schema.routes, retrieved.notify.on_complete.routes)
         self.assertEqual(retrieved.notify.on_success, None)
         self.assertEqual(retrieved.notify.on_failure, None)
 
@@ -106,7 +106,7 @@ class LiveActionModelTest(DbTestCase):
         self.assertEqual(notify_sub_schema.message,
                          retrieved.notify.on_success.message)
         self.assertDictEqual(notify_sub_schema.data, retrieved.notify.on_success.data)
-        self.assertListEqual(notify_sub_schema.channels, retrieved.notify.on_success.channels)
+        self.assertListEqual(notify_sub_schema.routes, retrieved.notify.on_success.routes)
         self.assertEqual(retrieved.notify.on_failure, None)
         self.assertEqual(retrieved.notify.on_complete, None)