+++ st2tests/st2tests/config.py
@@ -66,6 +66,7 @@ def _override_db_opts():
 def _override_common_opts():
     packs_base_path = get_fixtures_base_path()
     CONF.set_override(name='system_packs_base_path', override=packs_base_path, group='content')
+    CONF.set_override(name='packs_base_paths', override=packs_base_path, group='content')
     CONF.set_override(name='api_url', override='http://localhost', group='auth')
     CONF.set_override(name='admin_users', override=['admin_user'], group='system')
     CONF.set_override(name='mask_secrets', override=True, group='log')
@@ -108,7 +109,9 @@ def _register_api_opts():
 
     messaging_opts = [
         cfg.StrOpt('url', default='amqp://guest:guest@localhost:5672//',
-                   help='URL of the messaging server.')
+                   help='URL of the messaging server.'),
+        cfg.ListOpt('cluster_urls', default=[],
+                    help='URL of all the nodes in a messaging service cluster.')
     ]
     _register_opts(messaging_opts, group='messaging')
 
@@ -118,7 +121,10 @@ def _register_api_opts():
                    help='Location of the script on the remote filesystem.'),
         cfg.BoolOpt('allow_partial_failure',
                     default=False,
-                    help='How partial success of actions run on multiple nodes should be treated.')
+                    help='How partial success of actions run on multiple nodes should be treated.'),
+        cfg.BoolOpt('use_ssh_config',
+                    default=False,
+                    help='Use the .ssh/config file. Useful to override ports etc.')
     ]
     _register_opts(ssh_runner_opts, group='ssh_runner')
 
@@ -155,10 +161,13 @@ def _register_action_sensor_opts():
 
 def _register_mistral_opts():
     mistral_opts = [
-        cfg.StrOpt('v2_base_url', default='http://localhost:8989/v2',
-                   help='Mistral v2 API server root endpoint.'),
-        cfg.IntOpt('max_attempts', default=2),
-        cfg.IntOpt('retry_wait', default=1)
+        cfg.StrOpt('v2_base_url', default='http://localhost:8989/v2', help='v2 API root endpoint.'),
+        cfg.IntOpt('max_attempts', default=2, help='Max attempts to reconnect.'),
+        cfg.IntOpt('retry_wait', default=1, help='Seconds to wait before reconnecting.'),
+        cfg.StrOpt('keystone_username', default=None, help='Username for authentication.'),
+        cfg.StrOpt('keystone_password', default=None, help='Password for authentication.'),
+        cfg.StrOpt('keystone_project_name', default=None, help='OpenStack project scope.'),
+        cfg.StrOpt('keystone_auth_url', default=None, help='Auth endpoint for Keystone.')
     ]
     _register_opts(mistral_opts, group='mistral')