+++ CHANGELOG.rst
@@ -7,6 +7,61 @@ in development
 * Add new OpenStack Keystone authentication backend.
   [Itxaka Serrano]
 * Information about parent workflow is now a dict in child's context field. (improvement)
+* Fix a bug when some runner parameter default values where not overridden when a
+  falsey value was used in the action metadata parameter override (e.g. False, 0).
+  [Eugen C.]
+* Correctly return 404 if user requests an invalid path which partially maps to an existing
+  path. (bug-fix)
+* Add support for restarting sensors which exit with a non-zero status code to
+  the sensor container. Sensor container will now automatically try to restart
+  (up to 2 times) sensor processes which die with a non-zero status code. (improvement)
+* Support for RabbitMQ cluster. StackStorm works with a RabbitMQ cluster and switches
+  nodes on failover. (feature)
+* Add index to the ActionExecution model to speed up query. (improvement)
+* Fix sort key in the ActionExecution API controller. (bug-fix)
+* Introduce a Paramiko SSH runner that uses eventlets to run scripts or commands in parallel. (improvement) (experimental)
+* Add action parameters validation to Mistral workflow on invocation. (improvement)
+* Fix key name for error message in liveaction result. (bug-fix)
+* Fix 500 API response when rule with no pack info is supplied. (bug-fix)
+* Fix bug in trigger-instance re-emit (extra kwargs passed to manager is now handled). (bug-fix)
+* Rename notification "channels" to "routes". (improvement)
+* Make sure auth hook and middleware returns JSON and "Content-Type: application/json" header
+  in every response. (improvement, bug-fix)
+* Fix bug in triggers emitted on key value pair changes and sensor spawn/exit. When
+  dispatching those triggers, the reference used didn't contain the pack names
+  which meant it was invalid and lookups in the rules engine would fail. (bug-fix)
+* Allow user to include files which are written on disk inside the action create API payload.
+  (new feature)
+* Allow user to retrieve content of a file inside a pack by using the new
+  ``/packs/views/files/`` API endpoint. (new feature)
+* Handle sudo in paramiko remote script runner. (bug-fix)
+* Turn on paramiko ssh runner as the default ssh runner in prod configuration.
+  To switch to fabric runner, set ``use_paramiko_ssh_runner`` to false in st2.conf. (improvement)
+* Add OpenStack Keystone authentication configuration for Mistral. (improvement)
+* Abiltiy to add trace tag to TriggerInstance from Sensor. (feature)
+* Ability to view trace in CLI with list and get commands. (feature)
+* Add ability to add trace tag to ``st2 run`` CLI command. (feature)
+* Add ability to specify trace id in ``st2 run`` CLI command. (feature)
+* Update ``st2ctl`` to correctly start ``st2web`` even if even if Mistral is no installed.
+  (bug-fix, improvement)
+* Add X-Request-ID header to all API calls for easier debugging. (improvement)
+* Add new CLI commands for disabling and enabling content pack resources
+  (``{sensor,action,rule} {enable, disable} <ref or id>``) (feature)
+
+0.12.2 - August 11, 2015.
+-------------------------
+
+* Support local ssh config file in remote runners. (feature)
+* Changes to htpasswd file used in `flat_file` auth backend do not require
+  a restart of st2auth and consequently StackStorm. (feature)
+
+0.12.1 - July 31, 2015
+----------------------
+
+* Un-registering a pack also removes ``rules`` and ``action aliases`` from the pack. (bug-fix)
+* Disable parallel SSH in fabric runner which causes issues with eventlets. (bug-fix)
+* Fix executions stuck in ``running`` state if runner container throws exception. (bug-fix)
+* Fix cases where liveaction result in dict are escaped and passed to Mistral. (bug-fix)
 
 0.12.0 - July 20, 2015
 ----------------------
@@ -48,6 +103,18 @@ in development
 * Move /exp/actionalias/ and /exp/aliasexecution to /v1/actionalias/ and /v1/aliasexecution/
   respectively. (upgrade)
 * Display friendly message for error in parameters validation on action execution. (improvement)
+
+0.11.6 - July 2, 2015
+---------------------
+
+* Update all the code to handle all the datetime objects internally in UTC. (improvement, bug-fix)
+
+0.11.5 - July 1, 2015
+---------------------
+
+* Fix a bug where ``end_timestamp`` is not captured for Mistral workflow executions (bug-fix)
+* Fix a bug where the CLI failed to display Mistral workflow that errored (bug-fix)
+* Fix a bug where the published variables is not captured in the Mistral workflow result (bug-fix)
 
 0.11.4 - June 30, 2015
 ----------------------